Content Update
Brought in 3 images to the file, each one for important people or
machines mentioned within the essay.
Changed href’s for several items, as well as changed the title of the
web page and added an id to it for any needed reference.
Created a note for the side of the page with inline styles and a font
awesome icon.
Commented out several lines of code to get rid of on-screen image posts
as well as the nav link1 item.
Updated a comment to be more informative.
Inserted essay content that has been completed so far.
Added another txt div to act as a division between the essay and image
posts, if including them.
Added several image sources.
Changed a css style to an hsl value.
Created to txt files - one to credit all essay sources, and the other
for all image sources.